edit: rename plural variables
