+++ src/quiz/entities/user-question-response.entity.ts
@@ -9,7 +9,7 @@ import {
 } from 'typeorm';
 import { QuizAttempt } from './quiz-attempt.entity';
 import { Question } from '../../question/entities/question.entity';
-import { Answer } from '../../answer/entities/answer.entity';
+import { Answer } from '../../answer/entities/answers.entity';
 
 @Entity('user_question_responses')
 export class UserQuestionResponse {
